【修改描述】流控方案重构，增加可定位性 【DTS/CR/AR单号】 AR20240127801698 【合入版本】BeiMing 24.2.T1.B001,
Match-id-725215d6fad5d339fba6b281ea9514bb5acf19a1

+++ src/net/net_channel_mgr.cpp
@@ -59,7 +59,7 @@ BResult NetChannelMgr::AddChannel(NetNode dstNid, ChannelPtr &ch, uint8_t plane)
     }
     mChannelMgr.insert(std::make_pair(dstNid.whole, chInfo));
     NET_LOG_INFO("Add channel success, dstNid:" << dstNid.nid << ", pid:" << dstNid.pid << ", channel:" << ch->Id() <<
-        ", plane:" << plane << ".");
+        ", plane:" << static_cast<uint32_t>(plane) << ".");
     return BIO_OK;
 }
